move functions into base.py
